feat: delete author data

+++ src/app/api/authors/[id]/route.js
@@ -1,4 +1,4 @@
-import { createErrorRes, createSuccessRes } from "@/lib/dto/res-dto"
+import { createSuccessRes } from "@/lib/dto/res-dto"
 import { authorServerSchema } from "@/lib/schemas/author/author-server-schema"
 import AuthorService from "@/lib/services/author-service"
 import { generateErrorHttpRes } from "@/lib/utils/http"
@@ -30,15 +30,19 @@ export async function PUT(req, { params }) {
 export async function DELETE(req, { params }) {
   try {
     const { id } = await params
-    
-    await AuthorService.delete({id})
+
+    const author = await AuthorService.delete({ id: parseInt(id)})
 
     return NextResponse.json(
-      createSuccessRes({ message: 'Author successfully deleted, id: ' + id })
+      createSuccessRes({
+        message: 'Author successfully deleted, id: ' + id,
+        data: author
+      })
     )
+
   } catch (err) {
     console.error(err)
-
+    
     const httpErr = generateErrorHttpRes(err)
     return NextResponse.json(httpErr.errRes, { status: httpErr.status })
   }
